Lots of crap. Starting to move from manual AST walker in each node, to the visitor pattern. I use mixin templates in D to signficantly cut down the lines of hand-written boilerplate per node. So far, 'first pass' of compilation is almost done. Next up is the passes for instruction selection, label assignment, and code generation.

+++ src/wiz/parse/scanner.d
@@ -35,7 +35,7 @@ private enum State
 class Scanner
 {
     private char terminator;
-    private uint position;
+    private size_t position;
     private uint commentLine;
     private char[] buffer;
     private string text;
